Move error utils to external
