+++ src/input-parser/yaml/yaml-validator.ts
@@ -1,6 +1,6 @@
 import { Document, parseAllDocuments, YAMLParseError } from "yaml";
 
-import { formatParseError } from "../shared/utils/error-utils";
+import { formatParseError } from "../../utils/error/error-utils";
 import { isDocumentValid } from "./utils/yaml-utils";
 
 const parseYaml = (content: string): boolean => {
